refactoring - all tests passing
