+++ source/DasBlog.Tests/FunctionalTests/ComponentTests/SampleComponentTests.cs
@@ -1,0 +1,63 @@
+using System;
+using DasBlog.Tests.Support.Interfaces;
+using Constants = DasBlog.Tests.Support.Common.Constants;
+using Xunit;
+using Xunit.Abstractions;
+using newtelligence.DasBlog.Runtime;
+
+/*
+ * THERE IS NO logging to the console - in the debugger the log appears in the detail of the test results
+ * when run from the console the log appears in a log file assuming you provide the correct command line
+ * dotnet xunit cli is required to get console output
+ * If I do "dotnet xunit -diagnostics" it barfs with a reference to a missing Microsoft.Extensions.Options
+ */
+namespace DasBlog.Tests.FunctionalTests.ComponentTests
+{
+	[Collection(Constants.TestInfrastructureUsersCollection)]
+	public class SampleComponentTests : IClassFixture<ComponentTestPlatform>, IDisposable
+	{
+
+		private ComponentTestPlatform platform;
+		private ITestOutputHelper testOutputHelper;
+		private IDasBlogSandbox dasBlogSandbox;
+		public SampleComponentTests(ITestOutputHelper testOutputHelper, ComponentTestPlatform componentTestPlatform)
+		{
+					// logged messagees appear in the detail pane of Rider's test runner for
+					// Running a successful test
+					// Debugging a successful test
+					// Running a failed test
+					// Debugging a failed test
+					// logger under for Run and Debug
+			// I want to get hold of the xunit test results when running from the command line.
+			// "dotnet test" isn't providing them as far as I can see from my limited investigations
+			// Even after specifying RuntimeFramework as 2.1.4 in the csproj I got
+			// "dotnet xunit" should be more fruitful but fails with the following issue:
+			//   Could not load file or assembly 'Microsoft.Extensions.Options, Version=2.1.0.0,
+			// There is a nuget package for M.E.O 2.1.1 but the install fails and is rolled back
+			// There are no other v2 nuget packages except a preview 2.2.
+			// "dotnet xunit" seems to be a no-go - did not investigate whether it solved
+			// the original problem of locating and using test results.
+			// 
+			// It turns out that the following is not a bad start:
+			// "dotnet test --logger trx;LogfileName=test_results.xml --results-directory ./test_results"
+			//			test_results.xml will appear in <proj>/source/DasBlog.Tests/FunctionalTests/test_results
+			componentTestPlatform.CompleteSetup(testOutputHelper);
+			this.platform = componentTestPlatform;
+			this.testOutputHelper = testOutputHelper;
+			dasBlogSandbox = platform.CreateSandbox(Constants.VanillaEnvironment);
+			dasBlogSandbox.Init();
+		}
+
+		[Fact]
+		[Trait(Constants.CategoryTraitType, Constants.ComponentTestTraitValue)]
+		public void SimpleTest()
+		{
+			_ = platform;
+			Assert.True(true);
+		}
+		public void Dispose()
+		{
+			dasBlogSandbox?.Terminate();
+		}
+	}
+}
