+++ source/DasBlog.Tests/FunctionalTests/TestInfrastructureTests/InfrastructureTestPlatform.cs
@@ -1,4 +1,5 @@
 using DasBlog.Tests.FunctionalTests.Common;
+using DasBlog.Tests.Support.Interfaces;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit.Abstractions;
 
@@ -9,6 +10,10 @@ namespace DasBlog.Tests.FunctionalTests.TestInfrastructureTests
 		protected override void InjectDependencies(IServiceCollection services)
 		{
 			// nothing to do
+		}
+		public IDasBlogSandbox CreateSandbox(string environment)
+		{
+			return ServiceProvider.GetService<IDasBlogSandboxFactory>().CreateSandbox(ServiceProvider, environment);
 		}
 
 		protected override void CompleteSetupLocal()
