+++ source/DasBlog.Tests/FunctionalTests/TestInfrastructureTests/InfrastructureTests.cs
@@ -18,6 +18,7 @@ namespace DasBlog.Tests.FunctionalTests.TestInfrastructureTests
 	/// <summary>
 	/// usage: dotnet test --logger trx;LogfileName=test_results.xml --results-directory ./test_results --filter Category=TestInfrastructureTest
 	/// </summary>
+	[Collection(Constants.TestInfrastructureUsersCollection)]
 	public class InfrastructureTests : IClassFixture<InfrastructureTestPlatform>
 	{
 		private readonly InfrastructureTestPlatform platform;
@@ -29,6 +30,7 @@ namespace DasBlog.Tests.FunctionalTests.TestInfrastructureTests
 		}
 
 		[Fact]
+		[Trait(Constants.CategoryTraitType, Constants.TestInfrastructureTestTraitValue)]
 		public void TestLogger()
 		{
 			var loggerFactory = platform.ServiceProvider.GetService<ILoggerFactory>();
@@ -37,7 +39,7 @@ namespace DasBlog.Tests.FunctionalTests.TestInfrastructureTests
 			Assert.True(true);
 		}
 		[Fact]
-		[Trait("Category", "TestInfrastructureTest")]
+		[Trait(Constants.CategoryTraitType, Constants.TestInfrastructureTestTraitValue)]
 		public void Test()
 		{
 			var scriptRunner = platform.ServiceProvider.GetService<IScriptRunner>();
@@ -48,21 +50,22 @@ namespace DasBlog.Tests.FunctionalTests.TestInfrastructureTests
 		}
 
 		[Fact]
-		[Trait("Category", "TestInfrastructureTest")]
+		[Trait(Constants.CategoryTraitType, Constants.TestInfrastructureTestTraitValue)]
 		public void Runner_WhenTimedOut_ThrowsExcption()
 		{
 			ScriptRunnerOptions opts = new ScriptRunnerOptions();
 			opts.ScriptDirectory = Path.Combine(Utils.GetProjectRootDirectory(), Constants.ScriptsRelativePath);
 			opts.ScriptTimeout = 1;		// one millisecond
+			opts.ScriptExitTimeout = 10;
 			var accessor = new OptionsAccessor<ScriptRunnerOptions> {Value = opts};
 			ILogger<ScriptRunner> logger =
 				platform.ServiceProvider.GetService<ILoggerFactory>().CreateLogger<ScriptRunner>();
 			ScriptRunner runner = new ScriptRunner(accessor, logger);
-			Assert.Throws<Exception>(() => runner.Run("TestScript.cmd", new Dictionary<string, string>()));
+			Assert.Throws<Exception>(() => runner.Run("TestScript.cmd", runner.DefaultEnv));
 		}
 
 		[Fact]
-		[Trait("Category", "TestInfrastructureTest")]
+		[Trait(Constants.CategoryTraitType, Constants.TestInfrastructureTestTraitValue)]
 		[Trait("Chosen", "1")]
 		public void DetectChangesScript_WhenCleanDirectory_ReturnsNothing()
 		{
