+++ source/DasBlog.Tests/Support/ScriptRunner.cs
@@ -58,7 +58,8 @@ namespace DasBlog.Tests.Support
 					}.Concat(arguments).ToArray());
 				psi.RedirectStandardOutput = true;
 				psi.RedirectStandardError = true;
-				
+				logger.LogInformation($"script timeout: {scriptTimeout}, script exit timeout {scriptExitTimeout}");
+
 				int exitCode = int.MaxValue;
 				using (var ps = Process.Start(psi))
 				{
@@ -69,13 +70,15 @@ namespace DasBlog.Tests.Support
 					var result = ps.WaitForExit(scriptTimeout);
 					exitCode = result ? ps.ExitCode : int.MaxValue - 1;
 				}
+				logger.LogInformation($"exit code: {exitCode}");
 
 				ThrowExceptionForBadExitCode(exitCode, scriptPathAndFileName, scriptTimeout, psi);
 				return (exitCode, output.ToArray(), errs.ToArray());
 			}
-			catch (Exception )
-			{;
-				throw;
+//			finally
+			catch (Exception e)
+			{
+				throw new Exception(e.Message, e);
 			}
 		}
