+++ source/DasBlog.Tests/Support/Scripts/TestScript.cmd
@@ -6,4 +6,6 @@ echo abc
 echo this is a mistake 1>&2
 rem # if we don't have something other than "echo" then the output is not caught by the caller
 set
-exit %errorlevel%
+set exitcode=%errorlevel%
+ping 168.192.0.255 -n 1 -w %1
+exit %exitcode%
